broke out classes into multiple files for readability.
