+++ python/task_manager/background_task_manager.py
@@ -16,318 +16,8 @@ import sgtk
 from sgtk.platform.qt import QtCore
 from sgtk import TankError
 
-class _BackgroundTask(object):
-    """
-    Container class for a single task.
-
-    A task is a Python callable (function/method/class) that takes some arguments, does some work and returns its
-    result.  Each task will be run in a thread and tasks will be executed in priority order.
-
-    For example:
-
-        def task_fetch_status():
-            return status_of_something()
-        ...
-        task_manager.add_task(task_fetch_status)
-        ...
-        def on_task_completion(task, group, result):
-            status = result
-            # do something with the status
-
-    Additionally, tasks can be chained together so that the output of one task can be passed directly to the input
-    of one or more downstream tasks.  To achieve this, the upstream task must returns it's result as a dictionary
-    and this dictionary is added to the named parameters of any downstream tasks by the task manager.  Care should
-    be taken when constructing these tasks that the result of one upstream task doesn't unintentionally overwrite 
-    any existing named parameters for a downstream task.
-
-    For example: 
-
-        def task_fetch_status():
-            return {"status":status_of_something()}
-        def task_do_something(status):
-            result = None
-            if status:
-                result = result_of_doing_something()
-            return result
-        ...
-        status_task_id = task_manager.add_task(task_fetch_status, priority=1)
-        work_task_id = task_manager.add_task(task_do_something, priority=2, upstream_task_ids = [status_task_id])
-        ...
-        def on_task_completion(task, group, result):
-            if task.id = work_task_id:
-                # do something with the result
-                ...
-
-    Upstream tasks can be fed into multiple down-stream tasks and the task priority can also be different so for 
-    example all status fetches could be set to happen before all do-somethings by setting the priority accordingly.
-    Down-stream tasks will also not start before it's upstream tasks have completed.
-    """
-    def __init__(self, task_id, cbl, group, priority, args, kwargs):
-        """
-        Construction.
-
-        :param task_id:     The unique id for this task
-        :param cbl:         Callable to execute to perform the task
-        :param group:       The group that this task belongs to
-        :param priority:    The priority this task should be run with
-        :param args:        Additional arguments that should be passed to func
-        :param kwargs:      Additional named arguments that should be passed to func
-        """
-        self._uid = task_id
-
-        self._cbl = cbl
-        self._args = args or []
-        self._kwargs = kwargs or {}
-
-        self._group = group
-        self._priority = priority
-
-    def __repr__(self):
-        """
-        Create a string representation of this instance
-        :returns:   A string representation of this instance
-        """
-        return "[%d, G:%s, P:%s] %s" % (self._uid, self._group, self._priority, self._cbl.__name__)
-
-    @property
-    def uid(self):
-        """
-        :returns:   The unique id of this task
-        """
-        return self._uid
-
-    @property
-    def group(self):
-        """
-        :returns:   The group this task belongs to
-        """
-        return self._group
-
-    @property
-    def priority(self):
-        """
-        :returns:   The priority for this task
-        """
-        return self._priority
-
-    def append_upstream_result(self, result):
-        """
-        Append the result from an upstream task to this tasks kwargs.  In order for the result to be appended 
-        to the task it must be a dictionary.  Each entry in the result dictionary is then added to the tasks
-        named parameters so care should be taken when building the tasks that named parameters for a downstream
-        task are not unintentionally overwritten by the result of an upstream task. 
-
-        :param result:  A dictionary containing the result from an upstream task.  If result is not a dictionary
-                        then it will be ignored.
-        """
-        if result and isinstance(result, dict):
-            self._kwargs = dict(self._kwargs.items() + result.items())
-
-    def run(self):
-        """
-        Perform this task
-
-        :returns:   The result of performing the task
-        """
-        return self._cbl(*self._args, **self._kwargs)
-
-class _WorkerThreadA(QtCore.QThread):
-    """
-    Asynchrounous worker thread that can run tasks in a separate thread.  This implementation
-    implements a custom run method that loops over tasks until asked to quit.
-    """
-    # Signal emitted when a task has completed successfully
-    task_completed = QtCore.Signal(object, object)# task, result
-    # Signal emitted when a task has failed
-    task_failed = QtCore.Signal(object, object, object)# task, message, stacktrace
-
-    def __init__(self, parent=None):
-        """
-        Construction
-
-        :param parent:  The parent QObject for this thread
-        """
-        QtCore.QThread.__init__(self, parent)
-
-        self._task = None
-        self._process_tasks = True
-        self._mutex = QtCore.QMutex()
-        self._wait_condition = QtCore.QWaitCondition()
-
-    def run_task(self, task):
-        """
-        Run the specified task
-
-        :param task:    The task to run
-        """
-        self._mutex.lock()
-        try:
-            self._task = task
-        finally:
-            self._mutex.unlock()
-        self._wait_condition.wakeAll()
-
-    def shut_down(self):
-        """
-        Shut down the thread and wait for it to exit before returning
-        """
-        self._mutex.lock()
-        try:
-            self._process_tasks = False
-        finally:
-            self._mutex.unlock()
-        self._wait_condition.wakeAll()
-        self.wait()
-
-    def run(self):
-        """
-        The main thread run function.  Loops over tasks until asked to exit.
-        """
-        while True:
-            # get the next task to process:
-            task_to_process = None
-            self._mutex.lock()
-            try:
-                while self._process_tasks and not task_to_process:
-                    task_to_process = self._task
-                    self._task = None
-                    if not task_to_process:
-                        # wait until we have something to do...
-                        self._wait_condition.wait(self._mutex)
-
-                if not self._process_tasks:
-                    # stop processing
-                    break
-            finally:
-                self._mutex.unlock()
-
-            # run the task:
-            try:
-                result = task_to_process.run()
-
-                self._mutex.lock()
-                try:
-                    if not self._process_tasks:
-                        break
-                    # emit the result (non-blocking):
-                    self.task_completed.emit(task_to_process, result)
-                finally:
-                    self._mutex.unlock()
-            except Exception, e:
-                # something went wrong so emit failed signal:
-                self._mutex.lock()
-                try:
-                    if not self._process_tasks:
-                        break
-                    tb = traceback.format_exc()
-                    # emit failed signal (non-blocking):
-                    self.task_failed.emit(task_to_process, str(e), tb)
-                finally:
-                    self._mutex.unlock()
-
-class _WorkerThreadB(QtCore.QThread):
-    """
-    Asynchrounous worker thread that can run tasks in a separate thread.  This implementation
-    uses a separate worker object that exists in the new thread and then uses signals to
-    communicate back and forth.
-
-    Note, this recipe exhibits odd behaviour in PyQt.  When initially created, the instance returned
-    in the assignment isn't always of type _WorkerThreadB!, e.g.:
-
-        thread = _WorkerThreadB()
-        assert isinstance(thread, _WorkerThreadB) # this should never assert!
-
-    Although this recipe is recommended in Qt as it is arguably a more 'correct' use of QThreads, it
-    is currently advised that the the overriden run recipe (_WorkerThreadA) be used instead whilst
-    Toolkit needs to support PyQt. 
-    """
-    class _Worker(QtCore.QObject):
-        """
-        Thread worker that just does work when requested.
-        """
-        # Signal emitted when a task has completed successfully
-        task_completed = QtCore.Signal(object, object)# task, result
-        # Signal emitted when a task has failed
-        task_failed = QtCore.Signal(object, object, object)# task, message, stacktrace
-
-        def __init__(self):
-            """
-            Construction
-            """
-            QtCore.QObject.__init__(self, None)
-
-        def do_task(self, task):
-            """
-            Run a single task.
-            """
-            try:
-                # run the task:
-                result = task.run()
-                # emit result:
-                self.task_completed.emit(task, result)
-            except Exception, e:
-                # something went wrong so emit failed signal:
-                tb = traceback.format_exc()
-                self.task_failed.emit(task, str(e), tb)
-
-    # Signal used to tell the worker that a task should be run
-    work = QtCore.Signal(object)# task
-    # Signal emitted when a task has completed successfully
-    task_completed = QtCore.Signal(object, object)# task, result
-    # Signal emitted when a task has failed
-    task_failed = QtCore.Signal(object, object, object)# task, message, stacktrace
-
-    def __init__(self, parent=None):
-        """
-        Construction
-
-        :param parent:  The parent QObject for this thread
-        """
-        QtCore.QThread.__init__(self, parent)
-
-        # create the worker instance:
-        self._worker = _WorkerThreadB._Worker()
-
-        # move the worker to the thread and then connect up the signals 
-        # that are used to communicate with it: 
-        self._worker.moveToThread(self)
-        self.work.connect(self._worker.do_task)
-        self._worker.task_failed.connect(self.task_failed)
-        self._worker.task_completed.connect(self.task_completed)
-
-    def run_task(self, task):
-        """
-        Run the specified task
-
-        :param task:    The task to run
-        """
-        # signal the worker to run the task
-        self.work.emit(task)
-
-    def shut_down(self):
-        """
-        Shut down the thread and wait for it to exit before returning
-        """
-        self.quit()
-        self.wait()
-        # the worker has now been moved back into the main thread so lets
-        # parent it to this thread so that it gets safely cleaned up
-        self._worker.setParent(self)
-        self._worker = None
-
-    def run(self):
-        """
-        Normally we wouldn't need to override the run method as by default it just runs the event
-        loop for the thread but due to a but in Qt pre-4.8 we have to do some extra stuff to make
-        sure everything gets cleaned up properly!
-        """
-        # run the event loop:
-        self.exec_()
-
-        # before we quit, we need to move the worker back to the main thread.  This is to work around
-        # issues with Qt pre-4.8 where any QObject.deleteLater's aren't executed on thread exit
-        # which would result in the worker not being cleaned up correctly!
-        self._worker.moveToThread(QtCore.QCoreApplication.instance().thread())
+from .background_task import BackgroundTask
+from .worker_thread import WorkerThread
 
 class BackgroundTaskManager(QtCore.QObject):
     """
@@ -459,7 +149,7 @@ class BackgroundTaskManager(QtCore.QObject):
         # create a new task instance:
         task_id = self._next_task_id
         self._next_task_id += 1
-        new_task = _BackgroundTask(task_id, cbl, group, priority, task_args, task_kwargs)
+        new_task = BackgroundTask(task_id, cbl, group, priority, task_args, task_kwargs)
 
         # add the task to the pending queue:
         self._pending_tasks_by_priority.setdefault(priority, []).append(new_task)
@@ -619,10 +309,10 @@ class BackgroundTaskManager(QtCore.QObject):
             return None
 
         # create a new worker thread - note, there are two different implementations of the WorkerThread class
-        # that use two different recipes.  Although _WorkerThreadB is arguably more correct it has some issues
-        # in PyQt so _WorkerThreadA is currently preferred - see the notes in the class above for further details
-        thread = _WorkerThreadA(self)
-        if not isinstance(thread, _WorkerThreadA):
+        # that use two different recipes.  Although WorkerThreadB is arguably more correct it has some issues
+        # in PyQt so WorkerThread is currently preferred - see the notes in the class above for further details
+        thread = WorkerThread(self)
+        if not isinstance(thread, WorkerThread):
             # for some reason (probably memory corruption somewhere else) I've occasionally seen the above
             # creation of a worker thread return another arbitrary object!  Added this in here so the code
             # will at least continue correctly and not do unexpected things!
